Add HDF5Log benchmark.
